// src/lib/api/axios/types.ts
export interface AddressInfo {
  id: number;
  name: string;
  email: string;
  group: boolean;
}